Arm backend: Move QuantArgs to its own file in _passes/
The is part of an effort to reduce the number of functions/classes in
utility files. This class has no reason being in a utility file when it
has a clear purpose on its own.
